fix: storybook render white page
