+++ frontend/web/stories/Index.stories.tsx
@@ -1,3 +1,4 @@
+import Authenticated from "@/layouts/AuthenticatedLayout";
 import Index from "@pages/task/Index";
 import type { Meta, StoryObj } from "@storybook/react";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
@@ -12,14 +13,40 @@ export const SuccessBehavior = {
       handlers: [
         http.get(`${baseUrl}/tags`, () => {
           return HttpResponse.json({
-            firstName: "Neil",
-            lastName: "Maverick",
+            tags: [
+              {
+                id: 1,
+                name: "tag1",
+              },
+              {
+                id: 2,
+                name: "tag2",
+              },
+            ],
           });
         }),
         http.get(`${baseUrl}/tasks`, () => {
           return HttpResponse.json({
-            firstName: "Neil",
-            lastName: "Maverick",
+            tasks: [
+              {
+                id: 1,
+                title: "task1",
+              },
+              {
+                id: 2,
+                title: "task2",
+              },
+            ],
+          });
+        }),
+        http.get(`${baseUrl}/get-user`, () => {
+          return HttpResponse.json({
+            auth: {
+              user: {
+                id: 1,
+                name: "test user",
+              },
+            },
           });
         }),
       ],
@@ -40,11 +67,15 @@ const meta: Meta<typeof Index> = {
   },
   decorators: [
     (Story) => (
-      <QueryClientProvider client={queryClient}>
-        <JotaiProvider>
-          <Story />
-        </JotaiProvider>
-      </QueryClientProvider>
+      <div className="h-screen">
+        <QueryClientProvider client={queryClient}>
+          <JotaiProvider>
+            <Authenticated>
+              <Story />
+            </Authenticated>
+          </JotaiProvider>
+        </QueryClientProvider>
+      </div>
     ),
   ],
 };
